% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmrse.R
\name{coerce.lmrse}
\alias{coerce.lmrse}
\title{Combine lmrse object into a results data.frame}
\usage{
coerce.lmrse(x)
}
\arguments{
\item{x}{an object of class \code{"lmrse"}}
}
\value{
\code{coerce.lmrse} returns a results \code{data.frame} with
coefficients, standard errors and p-values for the
covariables.
}
\description{
\code{coerce.lmrse} constructs a results \code{data.frame}
from an \code{"lmrse"} object, where coefficients, standard
errors and p-values for each covariable are placed in
consecutive columns.
}
\author{
James Staley \href{mailto:jrstaley95@gmail.com}{jrstaley95@gmail.com}
}
