% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmrse.R
\name{print.summary.lmrse}
\alias{print.summary.lmrse}
\title{Print summary lmrse}
\usage{
\method{print}{summary.lmrse}(x, ...)
}
\arguments{
\item{x}{an object of class "lmrse".}
}
\description{
print.summary method for class "lmrse".
}
\author{
James R Staley <jrstaley95@gmail.com>
}
