% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmrse.R
\name{robustseCpp}
\alias{robustseCpp}
\title{Robust SE (C++)}
\usage{
robustseCpp(y = y, x = x, cluster = cluster)
}
\arguments{
\item{y}{matrix of methylation.}

\item{x}{matrix of other covariates (including intercept).}

\item{cluster}{clustering variable.}
}
\value{
A matrix of robust standard errors where the rows are the CpGs and the columns are the covariates including the intercept.
}
\description{
robustseCpp fits cluster robust standard errors (using C++Eigen) for longitudinal analyses across CpG sites using standard linear regression.
}
\examples{
### Data
y <- rnorm(5000000)
y <- matrix(y, ncol=1000)
colnames(y) <- paste0("var",1:1000)
x <- rnorm(5000)
cluster <- rep(1:1000,5)
covar <- cbind(rbinom(5000,1,0.5),rnorm(5000))
X <- cbind(1,x,covar)
colnames(X) <- c("intercept", "X", "covar1", "covar2")

### Analyses
beta <- t((solve(crossprod(X))\%*\%t(X))\%*\%y)
se <- robustse(y=y, x=X, cluster=cluster)
}
\author{
James R Staley <js16174@bristol.ac.uk>
}
