% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{robustseEigen}
\alias{robustseEigen}
\title{robustseEigen}
\usage{
robustseEigen(es, xs, cs, xxs, covs)
}
\arguments{
\item{es}{a matrix of residuals.}

\item{xs}{a matrix of covariates.}

\item{cs}{a vector of clusters.}

\item{xxs}{a matrix of XtX.}

\item{covs}{an empty variance-covariance matrix.}
}
\value{
A matrix of robust standard errors where the rows are the CpGs and the columns are the covariates including the intercept.
}
\description{
robustseEigen fits cluster robust standard errors with C++Eigen across markers.
}
\author{
James R Staley <js16174@bristol.ac.uk>
}
