% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmrse.R
\name{sand_se}
\alias{sand_se}
\title{Sandwich SE}
\usage{
sand_se(model, cluster)
}
\arguments{
\item{model}{output from linear model.}

\item{cluster}{clustering variable.}
}
\value{
A vector of robust standard errors for the covariates including the intercept.
}
\description{
sand_se fits cluster robust standard errors using a sandwich estimator.
}
\author{
James R Staley <james.staley@bristol.ac.uk>
}
