% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmrse.R
\name{sandwich.se}
\alias{sandwich.se}
\title{Sandwich SE}
\usage{
sandwich.se(model, cluster)
}
\arguments{
\item{model}{output from linear model}

\item{cluster}{clustering variable}
}
\value{
\code{sandwich.se} returns a vector of robust standard errors for the
covariates including the intercept.
}
\description{
\code{sandwich.se} fits cluster robust standard errors using a
sandwich estimator.
}
\examples{
# Data
y <- rnorm(5000)
x <- rnorm(5000)
cluster <- rep(1:1000, 5)
c1 <- rbinom(5000, 1, 0.5)
c2 <- rnorm(5000)

# Analyses
model <- lm(y ~ x + c1 + c2)
se <- sandwich.se(model = model, cluster = cluster)

}
\author{
James Staley \href{mailto:jrstaley95@gmail.com}{jrstaley95@gmail.com}
}
