% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmrse.R
\name{coerce.lmrse}
\alias{coerce.lmrse}
\title{Combine lmrse object into a results data.frame}
\usage{
coerce.lmrse(x)
}
\arguments{
\item{x}{an object of class \code{"lmrse"}}
}
\description{
coerce method for class \code{"lmrse"}, where
coefficients, standard errors and p-values are
consecutive columns for each covariable in the model.
}
\author{
James Staley \href{mailto:jrstaley95@gmail.com}{jrstaley95@gmail.com}
}
