% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmrse.R
\name{coerce.lmrse}
\alias{coerce.lmrse}
\title{Combine lmrse object into a results data.frame}
\usage{
coerce.lmrse(x)
}
\arguments{
\item{x}{an object of class "lmrse".}
}
\description{
coerce method for class "lmrse", where coefficients, standard errors and p-values are consecutive columns for each covariable in the model.
}
\author{
James R Staley <jrstaley95@gmail.com>
}
