% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmrse.R
\name{lmrse}
\alias{lmrse}
\alias{lmrse-package}
\title{Linear regression model with cluster robust standard errors}
\usage{
lmrse(formula, cluster, data = NULL)
}
\arguments{
\item{formula}{containing the marker matrix as the response and the exposure
and covariates as the dependent terms}

\item{cluster}{clustering variable}

\item{data}{an optional \code{data.frame} which contains the covariates specified
in the formula}
}
\value{
List of coefficients, SE and p-values matrices:

\item{coef}{a matrix of regression coefficients}

\item{se}{a matrix of standard errors}

\item{p}{a matrix of p-values}
}
\description{
\code{lmrse} fits a linear regression model with cluster robust
standard errors for all markers.
}
\examples{
# Data
y <- rnorm(5000000)
y <- matrix(y, ncol = 1000)
colnames(y) <- paste0("var", 1:1000)
x <- rnorm(5000)
cluster <- rep(1:1000, 5)
c1 <- rbinom(5000, 1, 0.5)
c2 <- rnorm(5000)

# Analyses
res <- lmrse(y ~ x + c1 + c2, cluster = cluster)

}
\author{
James Staley \href{mailto:jrstaley95@gmail.com}{jrstaley95@gmail.com}
}
