% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmrse.R
\name{print.lmrse}
\alias{print.lmrse}
\title{Print lmrse}
\usage{
\method{print}{lmrse}(x, ...)
}
\arguments{
\item{x}{an object of class "lmrse".}
}
\description{
print method for class "lmrse".
}
\author{
James R Staley <james.staley@bristol.ac.uk>
}
