% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmrse.R
\name{print.lmrse}
\alias{print.lmrse}
\title{Print lmrse}
\usage{
\method{print}{lmrse}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"lmrse"}}
}
\description{
print method for class \code{"lmrse"}.
}
\author{
James Staley \href{mailto:jrstaley95@gmail.com}{jrstaley95@gmail.com}
}
