% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmrse.R
\name{summary.lmrse}
\alias{summary.lmrse}
\title{Summarizing lmrse}
\usage{
\method{summary}{lmrse}(x, ...)
}
\arguments{
\item{x}{an object of class "lmrse".}
}
\description{
summary method for class "lmrse".
}
\author{
James R Staley <jrstaley95@gmail.com>
}
